server/requirements.txt
arff==0.9
Flask==1.1.2
gunicorn==20.0.4
javabridge==1.0.19
jvm==0.4.0b3
librosa==0.8.0
pipenv==2020.8.13
python-weka-wrapper3==0.1.14
Flask-Cors==3.0.9


